// src/app/routes.enum.ts
export enum RoutesEnum {
  login = 'login',
  register = 'register',
  registerSecretKey = 'registerSecretKey',
  portal = 'portal',

  main = 'main',
  mainOverview = 'overview',
  mainDocumentation = 'documentation',
  mainSupplyContracts = 'supply-contract',
  mainTariffs = 'tariff',
  dashboardTableCharts = 'table-charts',
  dashboardPieCharts = 'pie-charts',

  manage = 'manage',
  manageLocations = 'location',

  manageMeters = 'meter',
  manageMeterId = 'meterId',
  manageMeterGeneral = 'general',
  manageMeterReadings = 'meter-snapshot',

  manageBuildingType = 'building-type',

  manageBuildings = 'building',
  manageBuildingId = 'buildingId',
  manageBuildingGeneral = 'general',
  manageBuildingRooms = 'room',
  manageBuildingFloors = 'floor',

  manageFloors = 'floor',
  manageFloorId = 'floorId',
  manageFloorGeneral = 'general',
  manageFloorRooms = 'room',

  manageGroups = 'group',
  manageGroupId = 'groupId',
  manageGroupMembers = 'members',
  manageGroupRights = 'rights',
  manageGroupSubordinates = 'subordinates',

  manageRooms = 'room',
  manageRoomId = 'roomId',
  manageRoomGeneral = 'general',
  manageRoomEnvironmentalReadings = 'environmental-readings',

  manageUsers = 'user',
  manageInvitedUsers = 'invite',

  user = 'user',
  userProfile = 'profile',
  userGroups = 'groups',
  userSecurity = 'security',
  userActionHistory = 'action-history',
  userAdditional = 'additional',
}